backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
lancedb==0.17.0
duckdb==1.1.3
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
python-jose[cryptography]==3.3.0
pyarrow==18.1.0
